backend/requirements.txt
# This requirements file lists all third-party dependencies for this project.
Flask>=2.3,<4
Flask-WTF>=1.1,<2
Flask-Login>=0.6,<1
Flask-RESTful>=0.3,<1
gunicorn>=21,<23
google-cloud-ndb>=2.3,<3
google-cloud-storage>=2.16,<3
google-auth>=2.25,<3
six
